Fixed a bunch of warnings

+++ ixwebsocket/IXNetSystem.cpp
@@ -77,7 +77,7 @@ namespace ix
         tv.tv_sec = timeout / 1000;
         tv.tv_usec = (timeout % 1000) * 1000;
 
-        int ret = select(maxfd + 1, &readfds, &writefds, &errorfds, timeout != -1 ? &tv : NULL);
+        int ret = select(static_cast< int >( maxfd + 1 ), &readfds, &writefds, &errorfds, timeout != -1 ? &tv : NULL);
 
         if (ret < 0)
         {
